+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Use original slot elements when re-rendering a custom element
 * Set `data-he-rerendering` attribute while rerendering a custom element
 * Make custom element keep width & height while rerendering
+* Fix server-side `HTMLElement#removeAttribute()` not removing dataset attributes with dashes
 
 ## 2.3.16 (2024-02-15)